feat/ add handleCalculation function

+++ src/contexts/FunctionsContext.jsx
@@ -5,7 +5,7 @@ import { useState, useEffect, createContext, useContext } from "react";
 import Swal from 'sweetalert2'
 
 // import api
-import { getItems, createItem, deleteItem } from 'api/items'
+import { getItems, createItem, patchItem, deleteItem } from 'api/items'
 
 // import context
 import { ModalContext } from "./ModalContext";
@@ -23,7 +23,6 @@ export default function FunctionsContextProvider({ children }) {
     try {
       const items = await getItems()
       setCurrentItems(items.map(item => ({...item})))
-      console.log(currentItems)
     } catch (error) {
       console.error(error)
     } 
@@ -66,7 +65,7 @@ export default function FunctionsContextProvider({ children }) {
       const data = await createItem({
         name: inputNameValue,
         price: inputPriceValue,
-        amount: 0,
+        amount: 1,
       })
 
       setCurrentItems(prevItems => {
@@ -85,7 +84,8 @@ export default function FunctionsContextProvider({ children }) {
       Swal.fire({
         icon: 'success',
         text: '新增成功！',
-        timer: 1500
+        timer: 1000,
+        showConfirmButton: false
       })
 
       // clear data
@@ -99,10 +99,52 @@ export default function FunctionsContextProvider({ children }) {
     }
   }
 
+  ////////// patch item
+  // amount of item +1 or -1
+  const handleCalculation = async ({ id }, cal) => {
+    const currentItem = currentItems.find(item => item.id === id)
+    
+    if(currentItem.amount + cal < 0) return
+    
+    try {
+      await patchItem({
+        id,
+        amount: currentItem.amount + cal
+      })
+      
+      setCurrentItems(prevItems => {
+        return prevItems.map(item => {
+          if(item.id === id) {
+            return {
+              ...item,
+              amount: currentItem.amount + cal,
+            }
+          }
+          return item
+        })
+      })
+    } catch (error) {
+      console.error(error)
+    } 
+  }
+
   ////////// delete item
   const handleDelete = async ({ id }) => {
+
+    const result = await Swal.fire({
+      icon: "warning",
+      title: "確定要刪除嗎？",
+      showCancelButton: true,
+      confirmButtonColor: "#FF6600",
+      confirmButtonText: "確定",
+      cancelButtonText: "取消"
+    })
+
+    if(!result.isConfirmed) return
+
     try {
       await deleteItem(id)
+      
       setCurrentItems(prevItems => {
         return prevItems.filter(item => item.id !== id)
       })
@@ -128,6 +170,7 @@ export default function FunctionsContextProvider({ children }) {
         handlePriceChange,
         handleAddItem,
         handleDelete,
+        handleCalculation,
         getItemsAsync,
       }}
     >
